pyproject.toml
[tool.poetry]
name = "rates-derivative-grammar"
version = "0.1.0"
description = "Grammar to parse string representation of interest rate derivatives"
authors = ["PJCampi <pierrejean.campigotto@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
lark-parser = "^0.8"
bidict = "^0.19.0"
orderedset = "^2.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
tox = "3"
mypy = "^0.770"
pylint = "^2.4.4"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
